fix(lens): home the ProcessRss union on the shared terminal-workspace leaf
Move `ProcessRssSchema`/`ProcessRss` into `@kolu/terminal-workspace/schema` (the browser-safe leaf both packages already import) and import it from kolu-common/surface.ts and @kolu/padi/vocab.ts, deleting both byte-identical copies and both SEAL NOTEs so the honest three-way union has one declaration instead of a lockstep copy on each side of the seal.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-2, raised by lowy). Not pushed or merged.
