+++ packages/common/src/surface.ts
@@ -34,6 +34,10 @@ import {
   defineBuildInfo,
   surfaceAppSurfaceWith,
 } from "@kolu/surface-app/surface";
+// The honest three-way process-RSS union — composed below into `ProcessMemorySchema`.
+// Owned by the shared browser-safe leaf so both sides of the padi seal read one
+// declaration; its `ProcessRss` type is re-exported above for this module's importers.
+import { ProcessRssSchema } from "@kolu/terminal-workspace/schema";
 import type { TaskProgressSchema } from "anyagent/schemas";
 import { match } from "ts-pattern";
 import { z } from "zod";
@@ -66,6 +70,7 @@ export type {
   Foreground,
   OpenCodeInfo,
   PrResult,
+  ProcessRss,
   PrUnavailableSource,
   RestoreTarget,
   TerminalId,
@@ -240,36 +245,6 @@ export function applyPreferencesPatch(
     }),
   };
 }
-
-/** A single process's resident-set size as an HONEST three-way state, not a
- *  `number | null` that conflates "no process to measure" with "the read failed".
- *
- *   - `{ status: "ok", rssBytes }` — a live process answered.
- *   - `{ status: "absent" }` — there is no process to measure (down / not-yet-
- *     sampled). The expected "no value", not an error.
- *   - `{ status: "error" }` — the process was BELIEVED up yet its RSS read threw.
- *     A real anomaly the rail must surface distinctly from `absent`, so a failed
- *     read never renders identically to "no process" (the
- *     `caught-error-must-not-collapse-to-empty` rule — a server-side log is not a
- *     user surface).
- *
- *  A discriminated union (not an extra error flag beside a nullable number) so the
- *  three states are mutually exclusive by construction — there is no representable
- *  "error AND a stale rss".
- *
- *  SEAL NOTE: `@kolu/padi` declares a BYTE-IDENTICAL copy of this union
- *  (`ProcessRssSchema` in `@kolu/padi/surface`) — padi's `processMemory` cell
- *  needs it, and the package-boundary seal forbids padi importing `kolu-common`
- *  (proved in `server/src/seal.test.ts`) OR this heavily-imported module importing
- *  `@kolu/padi`. The two are kept identical on purpose; kolu-server's memory
- *  sampler folds padi's reading into this cell, so a shape drift breaks that fold
- *  at the typecheck (structural assignability) rather than silently. */
-export const ProcessRssSchema = z.discriminatedUnion("status", [
-  z.object({ status: z.literal("ok"), rssBytes: z.number() }),
-  z.object({ status: z.literal("absent") }),
-  z.object({ status: z.literal("error") }),
-]);
-export type ProcessRss = z.infer<typeof ProcessRssSchema>;
 
 /** Live process-memory readout for the chrome bar's identity rail — the RSS of the
  *  three server-side processes the rail names. The CLIENT's own JS-heap figure is
